Move CourseSelect and SubjectSelect to their own files
